Time: 0 ms (100.00%) | Memory: 10.5 MB (33.63%) - LeetSync

+++ 66-plus-one/plus-one.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     vector<int> plusOne(vector<int>& digits) 
     {
-        int i=digits.size()-1;
+        short int i=digits.size()-1;
         for(;i>=0;--i)
         {
             if(digits[i]==9)
